+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Offline playback support on Android and iOS
+- `SourceConfig.options` to enable configuring stream start position
+
+## [0.9.2] (2023-08-24)
+
+### Changed
+
+- Update Bitmovin's native iOS SDK version to `3.43.1`
+
+## [0.9.1] (2023-08-17)
+
+### Fixed
+
+- Android: Player sometimes plays sound but doesn't display video nor controls.
+
 ## [0.9.0] (2023-08-11)
 
 ### Added